Refatoração de Order Use casa, e distribuição de responsabilidade
