+++ src/main/resources/application.properties
@@ -3,7 +3,7 @@ spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.properties.hibernate.show_sql=true
 spring.datasource.initialization-mode=always
 ##Ibellinazzi alterar aqui depois
-spring.datasource.url=jdbc:mysql://localhost:'3306'/burguer51?useSSL=false&serverTimezone=UTC&allowPublicKeyRetrieval=true
+spring.datasource.url=jdbc:mysql://db:'3306'/burguer51?useSSL=false&serverTimezone=UTC&allowPublicKeyRetrieval=true
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 spring.datasource.username=GangstaBurguer
